+++ back/db/schema.rb
@@ -921,6 +921,8 @@ ActiveRecord::Schema.define(version: 2022_12_05_112729) do
     t.integer "min_budget", default: 0
     t.string "downvoting_method", default: "unlimited", null: false
     t.integer "downvoting_limited_max", default: 10
+    t.string "posting_method", default: "unlimited", null: false
+    t.integer "posting_limited_max", default: 1
     t.index ["project_id"], name: "index_phases_on_project_id"
   end
 
@@ -1056,6 +1058,8 @@ ActiveRecord::Schema.define(version: 2022_12_05_112729) do
     t.string "downvoting_method", default: "unlimited", null: false
     t.integer "downvoting_limited_max", default: 10
     t.boolean "include_all_areas", default: false, null: false
+    t.string "posting_method", default: "unlimited", null: false
+    t.integer "posting_limited_max", default: 1
     t.index ["slug"], name: "index_projects_on_slug", unique: true
   end