+++ back/spec/acceptance/projects_spec.rb
@@ -219,6 +219,8 @@ resource 'Projects' do
         parameter :visible_to, "Defines who can see the project, either #{Project::VISIBLE_TOS.join(',')}. Defaults to public.", required: false
         parameter :participation_method, "Only for continuous projects. Either #{ParticipationContext::PARTICIPATION_METHODS.join(',')}. Defaults to ideation.", required: false
         parameter :posting_enabled, 'Only for continuous projects. Can citizens post ideas in this project? Defaults to true', required: false
+        parameter :posting_method, "Only for continuous projects with posting enabled. How does posting work? Either #{ParticipationContext::POSTING_METHODS.join(',')}. Defaults to unlimited for ideation, and limited to one for native surveys.", required: false
+        parameter :posting_limited_max, 'Only for continuous projects with limited posting. Number of posts a citizen can perform in this project. Defaults to 1', required: false
         parameter :commenting_enabled, 'Only for continuous projects. Can citizens post comment in this project? Defaults to true', required: false
         parameter :voting_enabled, 'Only for continuous projects. Can citizens vote in this project? Defaults to true', required: false
         parameter :upvoting_method, "Only for continuous projects with voting enabled. How does voting work? Either #{ParticipationContext::VOTING_METHODS.join(',')}. Defaults to unlimited", required: false
@@ -324,6 +326,8 @@ resource 'Projects' do
         let(:participation_method) { project.participation_method }
         let(:presentation_mode) { 'map' }
         let(:posting_enabled) { project.posting_enabled }
+        let(:posting_method) { 'limited' }
+        let(:posting_limited_max) { 5 }
         let(:commenting_enabled) { project.commenting_enabled }
         let(:voting_enabled) { project.voting_enabled }
         let(:upvoting_method) { project.upvoting_method }
@@ -347,6 +351,8 @@ resource 'Projects' do
           expect(json_response.dig(:data, :attributes, :participation_method)).to eq participation_method
           expect(json_response.dig(:data, :attributes, :presentation_mode)).to eq presentation_mode
           expect(json_response.dig(:data, :attributes, :posting_enabled)).to eq posting_enabled
+          expect(json_response.dig(:data, :attributes, :posting_method)).to eq posting_method
+          expect(json_response.dig(:data, :attributes, :posting_limited_max)).to eq posting_limited_max
           expect(json_response.dig(:data, :attributes, :commenting_enabled)).to eq commenting_enabled
           expect(json_response.dig(:data, :attributes, :voting_enabled)).to eq voting_enabled
           expect(json_response.dig(:data, :attributes, :downvoting_enabled)).to be true
@@ -483,6 +489,8 @@ resource 'Projects' do
         parameter :visible_to, "Defines who can see the project, either #{Project::VISIBLE_TOS.join(',')}.", required: false
         parameter :participation_method, "Only for continuous projects. Either #{ParticipationContext::PARTICIPATION_METHODS.join(',')}.", required: false
         parameter :posting_enabled, 'Only for continuous projects. Can citizens post ideas in this project?', required: false
+        parameter :posting_method, "Only for continuous projects with posting enabled. How does posting work? Either #{ParticipationContext::POSTING_METHODS.join(',')}. Defaults to unlimited for ideation, and limited to one for native surveys.", required: false
+        parameter :posting_limited_max, 'Only for continuous projects with limited posting. Number of posts a citizen can perform in this project. Defaults to 1', required: false
         parameter :commenting_enabled, 'Only for continuous projects. Can citizens post comment in this project?', required: false
         parameter :voting_enabled, 'Only for continuous projects. Can citizens vote in this project?', required: false
         parameter :upvoting_method, "Only for continuous projects with voting enabled. How does voting work? Either #{ParticipationContext::VOTING_METHODS.join(',')}.", required: false
@@ -1274,6 +1282,8 @@ resource 'Projects' do
         parameter :visible_to, "Defines who can see the project, either #{Project::VISIBLE_TOS.join(',')}. Defaults to public.", required: false
         parameter :participation_method, "Only for continuous projects. Either #{ParticipationContext::PARTICIPATION_METHODS.join(',')}. Defaults to ideation.", required: false
         parameter :posting_enabled, 'Only for continuous projects. Can citizens post ideas in this project? Defaults to true', required: false
+        parameter :posting_method, "Only for continuous projects with posting enabled. How does posting work? Either #{ParticipationContext::POSTING_METHODS.join(',')}. Defaults to unlimited for ideation, and limited to one for native surveys.", required: false
+        parameter :posting_limited_max, 'Only for continuous projects with limited posting. Number of posts a citizen can perform in this project. Defaults to 1', required: false
         parameter :commenting_enabled, 'Only for continuous projects. Can citizens post comment in this project? Defaults to true', required: false
         parameter :voting_enabled, 'Only for continuous projects. Can citizens vote in this project? Defaults to true', required: false
         parameter :upvoting_method, "Only for continuous projects with voting enabled. How does voting work? Either #{ParticipationContext::VOTING_METHODS.join(',')}. Defaults to unlimited", required: false